refactor and clean up send-bitcoin-destination-screen and create DestinationField component
