+++ app/components/send-flow/index.ts
@@ -1,0 +1,3 @@
+import DestinationField from "./DestinationField"
+
+export { DestinationField }
